Add files via upload
The function you provided for us was incomplete and did not work by just copy and pasting. The function created created a variable for the argument called assignment2. However, in the code/expression itself it did not call that variable at all, so I had to change assignment to assignment 2. As well as changing someData to assignment2. To check if the function was written properly and as intended I used R's built in mean() to check my answer. It came out to be correct so now we have 2 functions that can find the mean of a vector.